intToLong

+++ frameworks/innerkitsimpl/accessor/src/exif_metadata.cpp
@@ -765,10 +765,10 @@ void ExifMetadata::GetFilterArea(const std::vector<std::string> &exifKeys,
 void ExifMetadata::FindRationalRanges(ExifContent *content,
     std::vector<std::pair<uint32_t, uint32_t>> &ranges, int index)
 {
-    for (int i = 0; i < static_cast<int>(content->entries[index]->components); i++) {
+    for (unsigned long i = 0; i < content->entries[index]->components; i++) {
         std::pair<uint32_t, uint32_t> range =
             std::make_pair(content->entries[index]->offset +
-            exif_format_get_size(content->entries[index]->format) * i, NUMERATOR_SIZE);
+                static_cast<unsigned long>(exif_format_get_size(content->entries[index]->format)) * i, NUMERATOR_SIZE);
         ranges.push_back(range);
     }
     return;
